backend/requirements.txt
fastapi
uvicorn
paddleocr
paddlepaddle==2.6.2
rapidfuzz
python-multipart
pandas
pdf2image
Pillow
numpy<2.0.0
protobuf==3.20.2
decorator==4.4.2
reportlab
python-docx
pyspellchecker
docx2pdf
google-generativeai
python-dotenv